# Modelos/Dto/NoticiasDto.cs
namespace ServiciosClinicaUNMSM.Modelos.Dto
{
    public class NoticiasDto
    {
        public int id_noticia { get; set; }
        public string url_noticia { get; set; }
        public string img_noticia { get; set; }
        public int estado { get; set; }
    }
}