# bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ServiciosClinicaUNMSM, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Autenticacion/Login": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioDto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/UsuarioDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Autenticacion/ValidarCorreo": {
      "get": {
        "tags": [
          "Autenticacion"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "correo_usuario",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UsuarioDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Autenticacion/CrearUsuario": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioDto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/UsuarioDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Inicio/ListaNoticias": {
      "get": {
        "tags": [
          "Inicio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/NoticiasDto"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "NoticiasDto": {
      "type": "object",
      "properties": {
        "id_noticia": {
          "format": "int32",
          "type": "integer"
        },
        "url_noticia": {
          "type": "string"
        },
        "img_noticia": {
          "type": "string"
        },
        "estado": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "UsuarioDto": {
      "type": "object",
      "properties": {
        "id_usuario": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "apellidoPaterno": {
          "type": "string"
        },
        "apellidoMaterno": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "contrasenia": {
          "type": "string"
        },
        "fechaCreacion": {
          "format": "date-time",
          "type": "string"
        },
        "fechaActualizacion": {
          "format": "date-time",
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "estado": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    }
  }
}